Extract code related to encoding from join.cpp (#2975)
* refactor code related to encoding

* new encoding.cpp

* rename and merge Encoding.h to encoding.h
